separating exceptions into their own module
